requirements.txt
pandas 
numpy
seaborn
matplotlib
scikit-learn
pymongo
# catboost
# xgboost
pyaml
dill
flask
certifi
python-dotenv
mlflow
dagshub
fastapi
uvicorn
python-multipart
-e .